+++ app/page.jsx
@@ -1,37 +1,45 @@
 import React from 'react';
 import axios from 'axios';
-
-// Config
-import { ENV, EPS, SECTIONS } from '@/lib/config';
-
-// Layout components
+import { ENV, EPS } from '@/lib/config';
 import MainPageContainer from '@/components/layout/MainPageContainer';
 import Footer from '@/components/layout/Footer';
-
-// Section components (server)
 import LandingSection from '@/components/sections/LandingSection';
 import SectionHeader from '@/components/items/SectionHeader';
 import PlatformsSection from '@/components/sections/PlatformsSection';
 import UpcomingShowsSection from '@/components/sections/UpcomingShowsSection';
 import MusicSection from '@/components/sections/MusicSection';
 import BandMembersSection from '@/components/sections/BandMembersSection';
-
-// Interactive components (client)
 import MenuButton from '@/components/interactive/MenuButton';
-
-// Styles
 import styles from '@/components/styles/module/Page.module.css';
 import textStyles from '@/components/styles/globals/text.module.css';
+
+const formatDate = (iso) => {
+  if (!iso) return 'TBA';
+  return new Intl.DateTimeFormat('en-US', { dateStyle: 'medium' }).format(new Date(iso));
+};
+
+const formatTime = (iso) => {
+  if (!iso) return '';
+  return new Intl.DateTimeFormat('en-US', { timeStyle: 'short' }).format(new Date(iso));
+};
 
 async function getShows() {
   try {
     const { data } = await axios.get(`${process.env.NEXT_PUBLIC_BASE_URL || 'http://localhost:3000'}/api/shows`, {
-      headers: {
-        'Cache-Control': 'no-store'
-      }
+      headers: { 'Cache-Control': 'no-store' }
     });
 
-    return data.success ? data.result : [];
+    if (!data.success) return [];
+
+    const now = new Date();
+    return data.result
+      .filter(show => show.dateTime && new Date(show.dateTime) >= now)
+      .sort((a, b) => new Date(a.dateTime) - new Date(b.dateTime))
+      .map(show => ({
+        ...show,
+        formattedDate: formatDate(show.dateTime),
+        formattedTime: formatTime(show.dateTime),
+      }));
   } catch (err) {
     console.error('Shows fetch failed:', err);
     return [];
