+++ components/sections/index.jsx
@@ -1,22 +1,18 @@
-// Sections (containers with multiple items)
 export { default as BandMembersSection } from './BandMembersSection';
 export { default as MusicSection } from './MusicSection';
 export { default as UpcomingShowsSection } from './UpcomingShowsSection';
 export { default as PlatformsSection } from './PlatformsSection';
 export { default as LandingSection } from './LandingSection';
 
-// Items (single units)
 export { default as SectionHeader } from '@/components/items/SectionHeader';
 export { default as MemberItem } from '@/components/items/MemberItem';
 export { default as AlbumItem } from '@/components/items/AlbumItem';
 export { default as PlatformItem } from '@/components/items/PlatformItem';
 export { default as ShowItem } from '@/components/items/ShowItem';
 
-// Interactive (client components)
 export { default as NavigationMenu } from '@/components/interactive/NavigationMenu';
 export { default as MenuButton } from '@/components/interactive/MenuButton';
 export { default as CopyButton } from '@/components/interactive/CopyButton';
 
-// Layout
 export { default as MainPageContainer } from '@/components/layout/MainPageContainer';
 export { default as Footer } from '@/components/layout/Footer';
